角色权限管理

+++ src/java/com/rtworld/pojo/Role.java
@@ -2,17 +2,18 @@ package com.rtworld.pojo;
 
 
 import java.io.Serializable;
-
-public class Role  implements Serializable {
-  private long roleid;
+public class  Role  implements Serializable {
+  private int  roleid;
   private String rolename;
   private String description;
   public long getRoleid() {
     return roleid;
   }
-  public void setRoleid(long roleid) {
+
+  public void setRoleid(int roleid) {
     this.roleid = roleid;
   }
+
   public String getRolename() {
     return rolename;
   }
